+++ dbt/adapters/firebolt/column.py
@@ -8,3 +8,7 @@ class FireboltColumn(Column):
     @classmethod
     def string_type(cls, size: int) -> str:
         return 'text'
+
+    @classmethod
+    def is_date(self) -> bool:
+        return self.dtype.lower() == 'date'