+++ dbt/include/firebolt/macros/materializations/models/incremental/is_incremental.sql
@@ -1,9 +1,11 @@
 {% macro is_incremental() %}
-    {#-- do not run introspective queries in parsing #}
+    {#- Return True if model is incremental, otherwise False. -#}
     {% if not execute %}
         {{ return(False) }}
     {% else %}
-        {% set relation = adapter.get_relation(this.database, this.schema, this.table) %}
+        {% set relation = adapter.get_relation(this.database,
+                                               this.schema,
+                                               this.table) %}
         {{ return(relation is not none
                   and relation.type == 'table'
                   and model.config.materialized == 'incremental'