+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v.1.1.0
+
+### Features
+
+- dbt-firebolt now supports the insert/overwrite incremental model.
+
+### Under the hood
+
+- dbt-firebolt now requires firebolt-sdk>=0.8
+
+### Breaking changes
+
+- dbt-firebolt uses firebolt-sdk's new auth token method and requires firebolt-sdk >= 0.8
+
 ## v.1.0.6
 
 ### Under the hood
@@ -23,7 +37,7 @@
 
 ### Features
 
-- dbt-firebolt now supports append-only incremental models.
+- dbt-firebolt now supports the append-only incremental model.
 - We now return better/more accurate responses from the cursor for logging/debugging.
 
 ### Under the hood